+++ noodles-sam/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-  * sam/header/record/value/map/tag: Implement `Copy` for `Tag<S>`.
+  * sam/header/record/value/map/tag: Implement `Copy` + `Display` for `Tag<S>`.
 
 ## 0.27.0 - 2023-04-27